moving FunctionConfig to different module
